refactor : add dog list module and ajust the navigation graph
