+++ app/build.gradle
@@ -98,6 +98,8 @@ dependencies {
 
     //core module
     implementation project(path: ':core')
+    implementation project(path: ':doglist')
+    implementation project(path: ':camera')
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
