+++ camera/src/main/java/com/kalex/dogescollection/camera/camera/utils/BitmapUtils.java
@@ -1,4 +1,4 @@
-package com.kalex.dogescollection.camera.camera.camera.utils;
+package com.kalex.dogescollection.camera.camera.utils;
 
 /*
  * Copyright 2020 Google LLC. All rights reserved.
@@ -34,6 +34,9 @@ import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 import androidx.camera.core.ExperimentalGetImage;
 import androidx.camera.core.ImageProxy;
+
+import com.kalex.dogescollection.camera.camera.camera.utils.FrameMetadata;
+
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
