+++ camera/src/main/res/navigation/camera_graph.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/camera_graph"
+    xmlns:tools="http://schemas.android.com/tools"
+    app:startDestination="@+id/CameraFragmentNav"
+    >
+
+    <dialog
+        android:id="@+id/CameraFragmentNav"
+        android:name="com.kalex.dogescollection.camera.camera.CameraFragment"
+        android:label="camera"
+        tools:layout="@layout/fragment_camera">
+        <action
+            android:id="@+id/action_CameraFragment_to_DogResultFragment"
+            app:destination="@id/dogResultFragment">
+            <argument
+                android:name="dog"
+                android:defaultValue="@null"
+                app:argType="com.kalex.dogescollection.core.model.data.alldogs.Dog"
+                app:nullable="true" />
+        </action>
+
+    </dialog>
+    <fragment
+        android:id="@+id/dogResultFragment"
+        android:name="com.kalex.dogescollection.camera.camera.DogResultFragment"
+        android:label="fragment_dog_result"
+        tools:layout="@layout/fragment_dog_result"
+        >
+        <action
+            android:id="@+id/action_DogResultFragment_to_CameraFragment"
+            app:destination="@id/CameraFragmentNav" />
+        <argument
+            android:name="dog"
+            android:defaultValue="@null"
+            app:argType="com.kalex.dogescollection.core.model.data.alldogs.Dog"
+            app:nullable="true" />
+
+        <action
+            android:id="@+id/action_dogResultFragment_to_DogListFragment"
+            app:destination="@id/doglist_graph" />
+
+    </fragment>
+
+</navigation>
