+++ core/src/main/res/layout/dog_detail_info_item.xml
