+++ core/src/main/res/layout/dog_detail_item.xml
