+++ doglist/build.gradle
@@ -8,7 +8,7 @@ plugins {
 }
 
 android {
-    namespace 'com.kalex.dogescollection.camera'
+    namespace 'com.kalex.dogescollection.doglist'
     compileSdk 33
 
     defaultConfig {
@@ -36,7 +36,6 @@ android {
         viewBinding true
     }
 }
-
 kapt {
     correctErrorTypes true
 }
@@ -47,36 +46,30 @@ dependencies {
     implementation "com.google.dagger:hilt-android:2.44"
     implementation 'androidx.preference:preference:1.2.0'
     implementation 'androidx.core:core-ktx:1.9.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
-
     kapt "com.google.dagger:hilt-compiler:2.44"
 
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.8.0'
 
-    //camera X
-    def camerax_version = "1.2.1"
-    implementation "androidx.camera:camera-core:${camerax_version}"
-    implementation "androidx.camera:camera-camera2:${camerax_version}"
-    implementation "androidx.camera:camera-lifecycle:${camerax_version}"
-
-    implementation "androidx.camera:camera-view:${camerax_version}"
-    implementation "androidx.camera:camera-extensions:${camerax_version}"
-
     //Coil-img handler
     implementation "io.coil-kt:coil:2.2.2"
 
-    //tensor flow
-    implementation 'org.tensorflow:tensorflow-lite:2.11.0'
-    implementation 'org.tensorflow:tensorflow-lite-support:0.4.3'
+    //Epoxy
+    implementation "com.airbnb.android:epoxy:5.1.1"
+    // Add the annotation processor if you are using Epoxy's annotations (recommended)
+    kapt  "com.airbnb.android:epoxy-processor:5.1.1"
 
     //core module
     implementation project(path: ':core')
-    //doglist module
-    implementation project(path: ':doglist')
+
+    implementation 'androidx.core:core-ktx:1.7.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.8.0'
+
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
