+++ doglist/src/main/res/layout/dog_list_item.xml
