+++ doglist/src/main/res/layout/vertical_linear_group.xml
