+++ doglist/src/main/res/navigation/doglist_graph.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/doglist_graph"
+    app:startDestination="@id/DogListFragment"
+    >
+    <fragment
+        android:id="@+id/DogListFragment"
+        android:name="com.kalex.dogescollection.doglist.presentation.ui.DogListFragment"
+        android:label="@string/dog_list_fragment_label"
+        tools:layout="@layout/dog_list_fragment"
+
+        >
+
+        <action
+            android:id="@+id/action_DogListFragment_to_DogListDetailFragment"
+            app:destination="@id/DogListDetailFragment">
+            <argument
+                android:name="dog"
+                android:defaultValue="@null"
+                app:argType="com.kalex.dogescollection.core.model.data.alldogs.Dog"
+                app:nullable="true" />
+        </action>
+    </fragment>
+
+    <fragment
+        android:id="@+id/DogListDetailFragment"
+        android:name="com.kalex.dogescollection.doglist.presentation.ui.DogDetailFragment"
+        android:label="@string/dog_detail_fragment_label"
+        tools:layout="@layout/dog_detail_fragment">
+
+        <action
+            android:id="@+id/action_DogListDetailFragment_to_DogListFragment"
+            app:destination="@id/DogListFragment" />
+
+        <argument
+            android:name="dog"
+            android:defaultValue="@null"
+            app:argType="com.kalex.dogescollection.core.model.data.alldogs.Dog"
+            app:nullable="true" />
+
+    </fragment>
+</navigation>
