+++ settings.gradle
@@ -20,3 +20,4 @@ rootProject.name = "DogesCollection"
 include ':app'
 include ':core'
 include ':camera'
+include ':doglist'
